Delete Schedule frontend created and working

+++ server/controllers/schedule.controller.js
@@ -2,33 +2,33 @@ import Schedule from "../models/schedule.model.js";
 import { createScheduleId } from "../services/schedule.service.js";
 
 export const getSchedule = async (req, res) => {
-    try {
-        const schedules = await Schedule.find();
-        res.status(200).json(schedules);
-    } catch (error) {
-        console.error("Error fetching schedules:", error);
-        res.status(500).json({ error: "Failed to fetch schedules" });
-    }
+  try {
+    const schedules = await Schedule.find();
+    res.status(200).json(schedules);
+  } catch (error) {
+    console.error("Error fetching schedules:", error);
+    res.status(500).json({ error: "Failed to fetch schedules" });
+  }
 };
 
 export const createSchedule = async (req, res) => {
-    const {start, end } = req.body;
-    try {
-        const id = await createScheduleId()
-        const newSchedule = new Schedule({id,start,end})
-        await newSchedule.save();
-        res.status(201).json({message:"Schedule created successfully"})
-    }catch(err){
-        res.status(500).json({message: "Error occured", error: err.message})
-    }
-}
+  const { start, end } = req.body;
+  try {
+    const id = await createScheduleId();
+    const newSchedule = new Schedule({ id, start, end });
+    await newSchedule.save();
+    res.status(201).json({ message: "Schedule created successfully" });
+  } catch (err) {
+    res.status(500).json({ message: "Error occured", error: err.message });
+  }
+};
 
-export const bookSchedule = async (req, res)=>{
-    const { id, cid } = req.body;
-    console.log(id)
+export const bookSchedule = async (req, res) => {
+  const { id, cid } = req.body;
+  console.log(id);
   try {
     const schedule = await Schedule.findOne({ id });
-    console.log('booking')
+    console.log("booking");
     if (!schedule) {
       return res.status(404).json({ message: "Schedule not found" });
     }
@@ -42,17 +42,23 @@ export const bookSchedule = async (req, res)=>{
   } catch (err) {
     res.status(500).json({ message: err.message });
   }
-}
+};
 
 export const deleteSchedule = async (req, res) => {
-  const {id} = req.body
+  const { id } = req.params;
   try {
-    const schedule = await Schedule.findOne({id})
-    if(schedule.booked === true){
-      res.status(501).json({message : "Already booked can't delete"})
+    const schedule = await Schedule.findOne({ id });
+    if (!schedule) {
+      return res.status(404).json({ message: "Schedule not found" });
     }
+
+    if (schedule.booked === true) {
+      return res.status(501).json({ message: "Already booked can't delete" });
+    }
+
     await schedule.deleteOne();
-  }catch(err){
-    res.status(500).json({error : err})
+    return res.status(200).json({ message: "Delete Successful" });
+  } catch (err) {
+    return res.status(500).json({ error: err.message || err });
   }
-}
+};
